Use Decimal for user balance

+++ core/db/models/user.py
@@ -2,7 +2,9 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime, Float, String, func
+from decimal import Decimal
+
+from sqlalchemy import BigInteger, DateTime, Numeric, String, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from core.db import Base
@@ -17,7 +19,7 @@ class User(Base):
         DateTime, default=func.now(), server_default=func.now()
     )
 
-    balance: Mapped[float] = mapped_column(Float, default=0)
+    balance: Mapped[Decimal] = mapped_column(Numeric(10, 2), default=0)
 
     vpn_configs: Mapped[list["VPN_Config"]] = relationship(  # noqa F821 # type: ignore
         "VPN_Config",
